🚑 What CI found that my machine could not
Four failures, and three of them were invisible locally for the same
reason in different clothes.

**The lint gate had never seen a Windows or macOS source.** golangci-lint
analyses the files the current GOOS selects, so a job named "strict, zero
suppressions" running on ubuntu has been skipping every _windows.go,
_darwin.go and _linux.go file this project has. Two dead conversions in
internal/preflight/fs_linux.go failed CI because the runner is Linux and my
machine is not; six more findings were sitting in the Windows sources,
already shipped. The job now lints windows and darwin too, so the name is
true.

**The second stub telling the same lie.** internal/cli/schema_test.go
answered a bare task on POST /api/v1/sync and an enveloped one on GET —
the identical false claim about the server that let B-031 ship green, in a
second file. Fixed there too; the guard now fails without the envelope.

**hx-boost swallows the anchor after all** (B-032). The help lot removed
three hx-boost="false" attributes as decoration, on the strength of a
browser scenario that passed — and it passed because it asserted the URL
rather than where the viewport ended up. CI's Chrome leaves the reader at
the top of an index of forty error codes. Every link carrying a fragment
now opts out, which is the remedy B-004 and B-013 already established, and
the scenario measures the scroll position.

I also pushed a code fix without re-running the full suite behind it, which
is how the schema stub reached CI at all. That is the process failure here,
not the defect.

+++ internal/preflight/fs_linux.go
@@ -66,9 +66,12 @@ func (systemInspector) Inspect(path string) (Filesystem, Space, error) {
 		return Filesystem{}, Space{}, fmt.Errorf("statfs %s: %w", target, err)
 	}
 
-	// st.Type is int64 on 64-bit and int32 on 32-bit Linux; the
-	// conversion is what makes both architectures build.
-	magic := int64(st.Type)
+	// st.Type and st.Bsize are int64 on every architecture this project
+	// releases for (linux/amd64 and linux/arm64). They are int32 on 32-bit
+	// Linux, where these two lines would stop compiling — which is the
+	// right way to find out, rather than a conversion that is dead on
+	// every build we actually ship and that `unconvert` flags as such.
+	magic := st.Type
 	name, known := linuxMagics[magic]
 	f := classify(name, "statfs(2) f_type")
 	if !known {
@@ -80,7 +83,7 @@ func (systemInspector) Inspect(path string) (Filesystem, Space, error) {
 	// f_bavail, not f_bfree: the blocks an ext filesystem reserves for
 	// root are free to the kernel and unavailable to the instance, and
 	// counting them would promise space a non-root process cannot have.
-	blockSize := int64(st.Bsize)
+	blockSize := st.Bsize
 	space := Space{
 		FreeBytes:  int64(st.Bavail) * blockSize, //nolint:gosec // G115: block counts of a real volume are orders of magnitude below the int64 range
 		TotalBytes: int64(st.Blocks) * blockSize, //nolint:gosec // G115: same
